misc/HtmlTestSuite/results/test-0938.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>select setCustomValidity</title>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

</head><body><select id="select_test"/>

<script>

test(() =&gt; {
  let elem = document.getElementById("select_test");
  assert_false(elem.validity.customError);
  elem.setCustomValidity("custom error");
  assert_true(elem.validity.customError);
}, "select setCustomValidity is correct")

</script>
</body></html>
